Moved parser-related classes from `cruise.umple.compiler` to `cruise.umple.parser`. Updated all references to the appropriate namespaces. Specifically named individual imports instead of generic `import cruise.umple.parser.*`.
Note: This commit builds successfully, but fails some of the tests.
