+++ cruise.umple/src/StateMachine.ump
@@ -47,6 +47,8 @@ class StateMachine
  */
 class State
 {
+  depend cruise.umple.parser.Position;
+
   isA Node;
   name;
   Boolean isConcurrent = { numberOfNestedStateMachines() > 1 }
@@ -78,7 +80,7 @@ class State
  */
 class Activity
 {
-  depend cruise.umple.compiler.Position;
+  depend cruise.umple.parser.Position;
   depend java.util.*;
 
   lazy Position position;
@@ -123,7 +125,7 @@ class Activity
  */
 class Transition
 {
-    
+  depend cruise.umple.parser.Position;
   Boolean isInternal = false;
   Boolean autoTransition = false;
 
@@ -145,7 +147,7 @@ class Transition
  */
 class Action
 {
-  depend cruise.umple.compiler.Position;
+  depend cruise.umple.parser.Position;
   
   lazy actionType;
   lazy Position position;
@@ -209,7 +211,7 @@ class Event
  */
 class Guard
 {
-  depend cruise.umple.compiler.Position;
+  depend cruise.umple.parser.Position;
   depend java.util.*;
   
   isA ConstraintTree;
