+++ cruise.umple/test/cruise/umple/compiler/CoupleTest.java
@@ -7,7 +7,7 @@
 
 */
 
-package cruise.umple.compiler;
+package cruise.umple.parser;
 
 import org.junit.*;
