+++ cruise.umple/test/cruise/umple/compiler/FilterTest.java
@@ -16,6 +16,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.junit.*;
+import cruise.umple.parser.ErrorTypeSingleton;
+import cruise.umple.parser.ParseResult;
 
 import cruise.umple.parser.analysis.RuleBasedParser;
 import cruise.umple.util.SampleFileWriter;
