+++ cruise.umple/test/cruise/umple/compiler/ParserTest.java
@@ -7,7 +7,7 @@
 
 */
 
-package cruise.umple.compiler;
+package cruise.umple.parser;
