+++ cruise.umple/test/cruise/umple/compiler/RuleInstanceTest.java
@@ -7,9 +7,10 @@
 
 */
 
-package cruise.umple.compiler;
+package cruise.umple.parser;
 
 import org.junit.*;
+
 
 public class RuleInstanceTest
 {
