+++ cruise.umple/test/cruise/umple/compiler/TextParserTest.java
@@ -7,7 +7,7 @@
 
 */
 
-package cruise.umple.compiler;
+package cruise.umple.parser;
 
 import org.junit.Before;
 import org.junit.Test;
