+++ cruise.umple/test/cruise/umple/compiler/UmpleUSEParserTest.java
@@ -14,6 +14,8 @@ import java.util.*;
 
 import org.junit.*;
 
+import cruise.umple.parser.ParseResult;
+import cruise.umple.parser.ErrorTypeSingleton;
 import cruise.umple.parser.analysis.RuleBasedParser;
 import cruise.umple.util.*;
 import java.io.*;
